+++ oauth2/CHANGELOG.md
@@ -2,8 +2,13 @@
 
 This file documents changes to the `workbench-oauth2` library, including notes on how to upgrade to new versions.
 
+## 0.7
+SBT dependency: `"org.broadinstitute.dsde.workbench" %% "workbench-oauth2" % "0.7-TRAVIS-REPLACE-ME"`
+Changed:
+- removed support for google oauth2, only azure b2c authentication is supported now in Terra UIs and swagger-ui
+
 ## 0.6
-SBT dependency: `"org.broadinstitute.dsde.workbench" %% "workbench-oauth2" % "d314413"`
+SBT dependency: `"org.broadinstitute.dsde.workbench" %% "workbench-oauth2" % "0.6-d314413"`
 Changed:
 - Updated `OpenIDProviderMetadata` class to include an optional `end_session_endpoint`